# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trentino Stage API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/reviews": {
      "get": {
        "summary": "Retrieve reviews of a profile",
        "description": "Fetches a paginated list of reviews for a specific profile",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "description": "Session token of the user making the request.",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "profileId",
            "required": true,
            "description": "ID of the profile whose reviews are being fetched.",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "description": "Starting index for pagination (default is 0).",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "description": "Ending index for pagination (default is 10).",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the reviews.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session token or required fields."
          },
          "403": {
            "description": "Access denied to the requested profile's reviews."
          },
          "404": {
            "description": "Profile not found."
          }
        }
      },
      "post": {
        "summary": "Create a new review",
        "description": "Allows a user to create a review for another profile. Only users can review companies and only companies can review users. Each profile can review another profile only once.",
        "tags": [
          "Reviews"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "reviewedProfile",
                  "title",
                  "review",
                  "rating"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Session token of the user creating the review."
                  },
                  "reviewedProfile": {
                    "type": "string",
                    "description": "ID of the profile being reviewed."
                  },
                  "title": {
                    "type": "string",
                    "description": "Title of the review."
                  },
                  "review": {
                    "type": "string",
                    "description": "Content of the review."
                  },
                  "rating": {
                    "type": "integer",
                    "description": "Rating for the profile (1-5)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review created successfully."
          },
          "403": {
            "description": "Invalid session token or unauthorized review."
          },
          "404": {
            "description": "Profile not found."
          },
          "405": {
            "description": "Missing parameters, invalid review type, or profile has already reviewed this profile."
          }
        }
      },
      "put": {
        "summary": "Update an existing review",
        "description": "Allows a user to update their own review. At least one field (title, review, or rating) must be provided for the update.",
        "tags": [
          "Reviews"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "reviewId"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Session token of the user updating the review."
                  },
                  "reviewId": {
                    "type": "string",
                    "description": "ID of the review to be updated."
                  },
                  "title": {
                    "type": "string",
                    "description": "New title for the review (optional)."
                  },
                  "review": {
                    "type": "string",
                    "description": "New content for the review (optional)."
                  },
                  "rating": {
                    "type": "integer",
                    "description": "New rating for the review (optional)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review updated successfully."
          },
          "401": {
            "description": "Missing or invalid session token or required fields."
          },
          "404": {
            "description": "Review not found."
          },
          "405": {
            "description": "No fields provided for update."
          }
        }
      }
    },
...
  },
...
}